+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 #### New Features
 
+- Added support for targeted delete-insert via the `overwrite_condition` parameter in `DataFrameWriter.save_as_table`
+
 #### Bug Fixes
 
 #### Improvements