+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.44.0 (YYYY-MM-DD)
+## 1.44.0 (2025-12-15)
 
 ### Snowpark Python API Updates